requirements.txt
python-dotenv
openai
gradio
pandas
jupyter
keras
# tensorflow
# tensorboard
numpy
matplotlib
seaborn
